Placed restart near level display
I have also made restart button in gameovermodal call the restartButton function
instead of history.go(), which i missed to do in last commit. I tried to rescale the canvas and add the restart button at the below. Unfortunately that was not easy since raw constants were used everywhere and that somehow disabled the jump control. So I had to add this button near level display. I tried to shrink the restart button to 55px but thats not working, due to image size(I think?).